Restructure "topics" feature folders.
