+++ coral/src/app/features/topics/components/list/TopicCard.tsx
@@ -9,7 +9,7 @@ import {
 } from "@aivenio/design-system";
 import classes from "src/app/features/topics/list/components/TopicCard.module.css";
 import { Topic } from "src/domain/topics";
-import { createTopicOverviewLink } from "src/app/features/topics/list/utils/create-topic-overview-link";
+import { createTopicOverviewLink } from "src/app/features/topics/utils/create-topic-overview-link";
 
 type TopicCardPros = Pick<
   Topic,
