+++ coral/src/app/features/topics/hooks/list/useGetTopics.test.tsx
@@ -1,6 +1,5 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { renderHook, waitFor } from "@testing-library/react";
-import { useGetTopics } from "src/app/features/topics/list/hooks/useGetTopics";
 import { ReactElement } from "react";
 import { server } from "src/services/api-mocks/server";
 import {
@@ -9,6 +8,7 @@ import {
   mockTopicGetRequest,
 } from "src/domain/topics/topics-api.msw";
 import { TopicEnv } from "src/domain/topics";
+import { useGetTopics } from "src/app/features/topics/hooks/list/useGetTopics";
 
 const queryClient = new QueryClient({
   defaultOptions: {
