--- coral/src/app/features/topics/list/index.ts
@@ -1,3 +1,0 @@
-import TopicList from "src/app/features/topics/list/components/TopicList";
-
-export { TopicList };
