--- coral/src/app/features/topics/select-env/index.ts
@@ -1,3 +1,0 @@
-import SelectEnv from "src/app/features/topics/select-env/components/SelectEnv";
-
-export { SelectEnv };
