--- coral/src/app/features/topics/select-team/index.ts
@@ -1,3 +1,0 @@
-import SelectTeam from "src/app/features/topics/select-team/components/SelectTeam";
-
-export { SelectTeam };
